requirements.txt
torch
torchvision
numpy
matplotlib
datasets
tqdm
pillow
pandas
altair